updating portfolio
